# Conflicts:
#	src/genie/genotype/genotype_payload.cc
#	src/genie/genotype/genotype_payload.h
#	test/genotype/encode-test.cc

+++ src/genie/contact/contact_parameters.h
@@ -131,7 +131,7 @@ class ContactParameters {
     void addSCMParam(ContactSubmatParameters&& scm_param);
     const SCMParamsDtype& getSCMParams() const;
 
-    uint32_t getNumBinEntries(uint8_t chr_ID, uint8_t interv_mult=1);
+    uint64_t getNumBinEntries(uint8_t chr_ID, uint8_t interv_mult=1);
     uint32_t getNumTiles(uint8_t chr_ID, uint8_t interv_mult=1);
 
     void write(core::Writer& writer) const;
